modified data saving
