+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "editor.inlineSuggest.showToolbar": "onHover"
+}
